requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Pillow==9.3.0
reportlab==3.6.12
PyPDF2==3.0.1